Demo game reorganization, refactoring, misc. - Extracted model and view logic from MainWindow, moved to dedicated subsystems - MainWindow now focused on UI logic only - Implemented GameModel, wraps the common Model and implements initialization and updates - Implemented GameView, wraps the common View and implements initialization and updates - Implemented UISystem, manages MainWindow and dev GUI - Revised game app initialization to ensure all subsystems are ready before game is launched - Removed leftover multithreading data structures from demo game and editor, to be revised later - Misc. fixes and refactoring
