+++ demo/game/VadonDemo/CMakeLists.txt
@@ -1,7 +1,9 @@
 add_subdirectory(Core)
+add_subdirectory(Model)
 add_subdirectory(Platform)
 add_subdirectory(Render)
 add_subdirectory(UI)
+add_subdirectory(View)
 
 target_sources(${VADONDEMO_CURRENT_TARGET}
     PRIVATE
