+++ demo/game/VadonDemo/Model/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(${VADONDEMO_CURRENT_TARGET}
+    PRIVATE
+        GameModel.hpp
+        GameModel.cpp
+   )
