+++ demo/game/VadonDemo/UI/CMakeLists.txt
@@ -2,4 +2,6 @@ target_sources(${VADONDEMO_CURRENT_TARGET}
     PRIVATE
         MainWindow.hpp
         MainWindow.cpp
+        UISystem.hpp
+        UISystem.cpp
    )
