+++ demo/game/VadonDemo/View/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(${VADONDEMO_CURRENT_TARGET}
+    PRIVATE
+        GameView.hpp
+        GameView.cpp
+   )
